perf: decompression rate improvement
This commit provides significant improvements to the existing decompression time but it has some caveats including more RAM usage, abandonment to the clean architechture etc. Keeping in mind the PubSub timeout limit it is estimated to have upper limit of 4-5GBs of decompression capacity at the moment hence its a temporary solution.